Feb 14 2017 Update
Added the ability to create Unity Database Assets for your user-defined
class (as opposed to storing the data in a SQL database).
Added the ability to store a Sprite/Image in your class.  Sprite/Image
is also stored in the database (sprite/image should be located in your
"Resources" directory).