Feb 28 2017 Update
Added several methods to the Class cs file to make it more accessible.
Added ability to search for an object in the database returning the
Class type (not just the base class type).
Added the ability to pull the entire database in, and return it as a
List.
Added class creation methods to clone from a Base of the class.