May 10 2017 Update 2
Added a check to see if the IntValue of a new Enum property was already
being used by something else.  If it does exist, the new property is not
added to the enum.

Added proper citing for the JobManager script to the appropriate author.

Added notation in the Crypto class to remind the user to modify the
starting values to something unique, specific to their game/application.